Add virtual touchscreen uinput driver and enable it on Beagleboard.
This fixes issues where Android apps exit unexpectedly. This is due
to application coding that relies on a touchscreen being present
in order to find certain layout and drawable resources. This can be
detected in logcat by noting messages that state that a resource
could not be found when on a system with no touchscreen.

Typically these apps can still be navigated with just keyboard
and trackball/mouse input anyway.